feat: widen price and tick size to u128 (Candid nat)
Widen the order-book `Price` key and `TickSize` from `u64` to `u128` (Candid
`nat`) so pairs whose per-whole-base price exceeds `u64` are representable.
`price`/`tick_size` Candid fields become `nat`; `lot_size` and `quantity` are
unchanged. Adds `Quantity::checked_mul_u128` and `u128`/`NonZeroU128` CBOR
helpers, threads `Nat ↔ u128` conversions through the canister boundary, and
widens the benchmarks accordingly.

DEFI-2848

+++ canister/src/dashboard/mod.rs
@@ -30,7 +30,7 @@ pub struct DashboardPair {
     pub book_id: u64,
     pub base_symbol: String,
     pub quote_symbol: String,
-    pub tick_size: u64,
+    pub tick_size: u128,
     pub lot_size: u64,
     pub bids_len: usize,
     pub asks_len: usize,
@@ -38,12 +38,12 @@ pub struct DashboardPair {
     pub resting_orders_len: usize,
     pub best_bid: Option<DashboardLevel>,
     pub best_ask: Option<DashboardLevel>,
-    pub spread: Option<u64>,
+    pub spread: Option<u128>,
     pub depth: DashboardDepth,
 }
 
 pub struct DashboardLevel {
-    pub price: u64,
+    pub price: u128,
     pub quantity: String,
 }
 
@@ -59,7 +59,7 @@ impl DashboardDepth {
 }
 
 pub struct DashboardDepthLevel {
-    pub price: u64,
+    pub price: u128,
     pub quantity: String,
     pub bar_width_percent: u8,
 }
