+++ canister/src/state/audit/tests.rs
@@ -373,7 +373,7 @@ fn should_replay_add_limit_order() {
         .with_limit_order(
             user_1(),
             Side::Buy,
-            Price::new(price * PRICE_SCALE),
+            Price::new(price as u128 * PRICE_SCALE),
             Quantity::from(quantity),
         );
     scenario.assert_replay_matches();
@@ -398,13 +398,13 @@ fn should_replay_matching() {
         .with_limit_order(
             buyer,
             Side::Buy,
-            Price::new(price * PRICE_SCALE),
+            Price::new(price as u128 * PRICE_SCALE),
             Quantity::from(quantity),
         );
     let (scenario, sell_id) = scenario.with_limit_order(
         seller,
         Side::Sell,
-        Price::new(price * PRICE_SCALE),
+        Price::new(price as u128 * PRICE_SCALE),
         Quantity::from(quantity),
     );
 
@@ -473,13 +473,13 @@ fn should_replay_matching_with_price_improvement() {
         .with_limit_order(
             seller,
             Side::Sell,
-            Price::new(maker_price * PRICE_SCALE),
+            Price::new(maker_price as u128 * PRICE_SCALE),
             Quantity::from(quantity),
         );
     let (scenario, buy_id) = scenario.with_limit_order(
         buyer,
         Side::Buy,
-        Price::new(taker_price * PRICE_SCALE),
+        Price::new(taker_price as u128 * PRICE_SCALE),
         Quantity::from(quantity),
     );
 
@@ -543,7 +543,7 @@ fn should_replay_cancel_pending_order() {
         .with_limit_order(
             user_1(),
             Side::Buy,
-            Price::new(price * PRICE_SCALE),
+            Price::new(price as u128 * PRICE_SCALE),
             Quantity::from(quantity),
         );
 
@@ -582,13 +582,13 @@ fn should_replay_cancel_partially_filled_order() {
         .with_limit_order(
             seller,
             Side::Sell,
-            Price::new(price * PRICE_SCALE),
+            Price::new(price as u128 * PRICE_SCALE),
             Quantity::from(quantity),
         );
     let (scenario, buy_id) = scenario.with_limit_order(
         buyer,
         Side::Buy,
-        Price::new(price * PRICE_SCALE),
+        Price::new(price as u128 * PRICE_SCALE),
         Quantity::from(3 * quantity),
     );
